IB: Added reboot if IB is not active
An extra step was added in the Powershell script after all VMs have
rebooted. A check for the modules will be made on each VM and if any is
missing, a reboot will be pefromed. We have 4 retries/VM. If after 4
retries IB is not up (modules are missing), then the test will be
aborted